+++ test/integration/api_query_test.lua
@@ -347,6 +347,7 @@ end
 
 function g.test_servers()
     local router = g.cluster:server('router')
+    local app_version = 'app_version_test_value'
 
     local resp = router:graphql({
         query = [[
@@ -366,6 +367,7 @@ function g.test_servers()
                         vshard_router { buckets_unreachable }
                         vshard_storage { buckets_active }
                         general {
+                            app_version
                             http_port
                             http_host
                             webui_prefix
@@ -390,7 +392,7 @@ function g.test_servers()
                 membership = {status = 'alive'},
                 vshard_router = {{ buckets_unreachable = 0 }},
                 vshard_storage = box.NULL,
-                general = { http_port = 8081, http_host = "0.0.0.0", webui_prefix = "" },
+                general = { app_version = app_version, http_port = 8081, http_host = "0.0.0.0", webui_prefix = "" },
             },
         }, {
             uri = 'localhost:13302',
@@ -406,7 +408,7 @@ function g.test_servers()
                 membership = {status = 'alive'},
                 vshard_router = box.NULL,
                 vshard_storage = {buckets_active = 3000},
-                general = { http_port = 8082, http_host = "0.0.0.0", webui_prefix = "" },
+                general = { app_version = app_version, http_port = 8082, http_host = "0.0.0.0", webui_prefix = "" },
             },
         }, {
             uri = 'localhost:13304',
@@ -422,7 +424,7 @@ function g.test_servers()
                 membership = {status = 'alive'},
                 vshard_router = box.NULL,
                 vshard_storage = {buckets_active = 3000},
-                general = { http_port = 8084, http_host = "0.0.0.0", webui_prefix = "" },
+                general = { app_version = app_version, http_port = 8084, http_host = "0.0.0.0", webui_prefix = "" },
             },
         }, {
             uri = 'localhost:13303',